moved AuthService to AuthStoreService
