+++ src/app/modules/group/pages/group-edit/group-edit.page.ts
@@ -99,9 +99,7 @@ export class GroupEditPage implements OnInit {
     });
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
