+++ src/app/modules/group/pages/group-list/group-list.page.ts
@@ -53,9 +53,7 @@ export class GroupListPage implements OnInit {
     this.getGroups();
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
