+++ src/app/modules/group/pages/group-members/group-members.page.ts
@@ -35,9 +35,7 @@ export class GroupMembersPage implements OnInit {
 
   ngOnInit() {}
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
 }
