+++ src/app/modules/group/pages/group-profile/group-profile.page.ts
@@ -86,9 +86,7 @@ export class GroupProfilePage implements OnInit {
       });
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
