+++ src/app/modules/group/pages/partners/partners.page.ts
@@ -57,9 +57,7 @@ export class PartnersPage implements OnInit {
       });
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
