+++ src/app/modules/group/pages/search/search.page.ts
@@ -56,9 +56,7 @@ export class SearchPage implements OnInit {
     this.getGroups();
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
