+++ src/app/modules/user/pages/friends/friends.page.ts
@@ -72,9 +72,7 @@ export class FriendsPage implements OnInit {
       });
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
