+++ src/app/modules/user/pages/user-groups/user-groups.page.ts
@@ -67,9 +67,7 @@ export class UserGroupsPage implements OnInit {
       });
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
