+++ src/app/modules/user/pages/user-profile/user-profile.page.ts
@@ -83,9 +83,7 @@ export class UserProfilePage implements OnInit {
       });
   }
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
