+++ src/app/modules/user/pages/user-settings/user-settings.page.ts
@@ -50,9 +50,7 @@ export class UserSettingsPage implements OnInit {
 
   ngOnInit() {}
 
-  ionViewWillEnter() {
-    this.menuService.onEnter();
-  }
+  ionViewWillEnter() {}
 
   ionViewWillLeave() {}
 }
